--- a/rust_snuba/benches/processors.rs
+++ b/rust_snuba/benches/processors.rs
@@ -16,13 +16,18 @@ use rust_arroyo::processing::{Callbacks, ConsumerState, RunError, StreamProcesso
 use rust_arroyo::types::{Partition, Topic};
 use rust_arroyo::utils::clock::SystemClock;
 use rust_snuba::{
-    ClickhouseConfig, ConsumerStrategyFactory, MessageProcessorConfig, StorageConfig, StatsDBackend,
+    ClickhouseConfig, ConsumerStrategyFactory, MessageProcessorConfig, StatsDBackend, StorageConfig,
 };
 use uuid::Uuid;
 
 fn main() {
     // this sends to nowhere, but because it's UDP we won't error.
-    configure_metrics(StatsDBackend::new("127.0.0.1", 8081, "snuba.consumer", Default::default()));
+    configure_metrics(StatsDBackend::new(
+        "127.0.0.1",
+        8081,
+        "snuba.consumer",
+        Default::default(),
+    ));
     divan::main();
 }