--- a/rust_snuba/src/lib.rs
+++ b/rust_snuba/src/lib.rs
@@ -23,6 +23,6 @@ fn rust_snuba(_py: Python<'_>, m: &PyModule) -> PyResult<()> {
 // plus a pyo3 specific crate as `cdylib`.
 pub use config::{ClickhouseConfig, MessageProcessorConfig, StorageConfig};
 pub use factory::ConsumerStrategyFactory;
+pub use metrics::statsd::StatsDBackend;
 pub use strategies::noop::Noop;
 pub use strategies::python::PythonTransformStep;
-pub use metrics::statsd::StatsDBackend;