re-attempt introduction of statsdproxy, failed previously at #5064. remove dbg!() statement that slows down the entire rust consumer. also fix up benchmarks to actually include statsd overhead.

main branch with benchmark fixes:

```

Timer precision: 41 ns
processors    fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ functions                │               │               │               │         │
│  ├─ 1       80.34 ms      │ 177.5 ms      │ 91.45 ms      │ 92.68 ms      │ 100     │ 100
│  │          62.23 Kitem/s │ 28.16 Kitem/s │ 54.67 Kitem/s │ 53.94 Kitem/s │         │
│  ├─ 4       51.08 ms      │ 110.4 ms      │ 57.13 ms      │ 59.04 ms      │ 100     │ 100
│  │          97.87 Kitem/s │ 45.28 Kitem/s │ 87.51 Kitem/s │ 84.68 Kitem/s │         │
│  ╰─ 16      29.58 ms      │ 51.14 ms      │ 32.56 ms      │ 34.68 ms      │ 100     │ 100
│             169 Kitem/s   │ 97.75 Kitem/s │ 153.5 Kitem/s │ 144.1 Kitem/s │         │
├─ profiles                 │               │               │               │         │
│  ├─ 1       65.33 ms      │ 95.38 ms      │ 69.42 ms      │ 71.36 ms      │ 100     │ 100
│  │          76.53 Kitem/s │ 52.42 Kitem/s │ 72.02 Kitem/s │ 70.06 Kitem/s │         │
│  ├─ 4       41.42 ms      │ 96.97 ms      │ 46.57 ms      │ 48 ms         │ 100     │ 100
│  │          120.7 Kitem/s │ 51.56 Kitem/s │ 107.3 Kitem/s │ 104.1 Kitem/s │         │
│  ╰─ 16      25.23 ms      │ 45.74 ms      │ 28.54 ms      │ 30.2 ms       │ 100     │ 100
│             198.1 Kitem/s │ 109.2 Kitem/s │ 175.1 Kitem/s │ 165.5 Kitem/s │         │
├─ querylog                 │               │               │               │         │
│  ├─ 1       167.6 ms      │ 311.9 ms      │ 184 ms        │ 188.6 ms      │ 100     │ 100
│  │          29.81 Kitem/s │ 16.02 Kitem/s │ 27.16 Kitem/s │ 26.51 Kitem/s │         │
│  ├─ 4       111.8 ms      │ 139.3 ms      │ 121 ms        │ 122.1 ms      │ 100     │ 100
│  │          44.7 Kitem/s  │ 35.88 Kitem/s │ 41.29 Kitem/s │ 40.91 Kitem/s │         │
│  ╰─ 16      66.76 ms      │ 137.9 ms      │ 74.17 ms      │ 78.17 ms      │ 100     │ 100
│             74.89 Kitem/s │ 36.23 Kitem/s │ 67.4 Kitem/s  │ 63.95 Kitem/s │         │
╰─ spans                    │               │               │               │         │
   ├─ 1       97.55 ms      │ 138.1 ms      │ 105.4 ms      │ 107.8 ms      │ 100     │ 100
   │          51.25 Kitem/s │ 36.19 Kitem/s │ 47.4 Kitem/s  │ 46.37 Kitem/s │         │
   ├─ 4       64.31 ms      │ 139.5 ms      │ 69.88 ms      │ 74.01 ms      │ 100     │ 100
   │          77.74 Kitem/s │ 35.82 Kitem/s │ 71.54 Kitem/s │ 67.55 Kitem/s │         │
   ╰─ 16      36.41 ms      │ 62.35 ms      │ 41.85 ms      │ 43.42 ms      │ 100     │ 100
              137.3 Kitem/s │ 80.18 Kitem/s │ 119.4 Kitem/s │ 115.1 Kitem/s │         │
```

this branch:

```
Timer precision: 41 ns [statsdproxy]
processors    fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ functions                │               │               │               │         │
│  ├─ 1       50.07 ms      │ 106.3 ms      │ 53.8 ms       │ 55.97 ms      │ 100     │ 100
│  │          99.85 Kitem/s │ 47 Kitem/s    │ 92.92 Kitem/s │ 89.32 Kitem/s │         │
│  ├─ 4       27.59 ms      │ 30.41 ms      │ 29.06 ms      │ 29.06 ms      │ 100     │ 100
│  │          181.2 Kitem/s │ 164.4 Kitem/s │ 172 Kitem/s   │ 172 Kitem/s   │         │
│  ╰─ 16      19.64 ms      │ 32.11 ms      │ 20.54 ms      │ 20.72 ms      │ 100     │ 100
│             254.4 Kitem/s │ 155.7 Kitem/s │ 243.4 Kitem/s │ 241.2 Kitem/s │         │
├─ profiles                 │               │               │               │         │
│  ├─ 1       41.35 ms      │ 72.7 ms       │ 43.74 ms      │ 45.5 ms       │ 100     │ 100
│  │          120.8 Kitem/s │ 68.77 Kitem/s │ 114.2 Kitem/s │ 109.8 Kitem/s │         │
│  ├─ 4       24.98 ms      │ 63.48 ms      │ 26.59 ms      │ 27.37 ms      │ 100     │ 100
│  │          200.1 Kitem/s │ 78.75 Kitem/s │ 188 Kitem/s   │ 182.6 Kitem/s │         │
│  ╰─ 16      18.28 ms      │ 40.68 ms      │ 19.53 ms      │ 20.64 ms      │ 100     │ 100
│             273.3 Kitem/s │ 122.8 Kitem/s │ 255.9 Kitem/s │ 242.1 Kitem/s │         │
├─ querylog                 │               │               │               │         │
│  ├─ 1       101.1 ms      │ 152.1 ms      │ 104.8 ms      │ 107.1 ms      │ 100     │ 100
│  │          49.45 Kitem/s │ 32.86 Kitem/s │ 47.68 Kitem/s │ 46.68 Kitem/s │         │
│  ├─ 4       53.53 ms      │ 70.61 ms      │ 57.42 ms      │ 58.19 ms      │ 100     │ 100
│  │          93.38 Kitem/s │ 70.8 Kitem/s  │ 87.07 Kitem/s │ 85.91 Kitem/s │         │
│  ╰─ 16      41.95 ms      │ 49.09 ms      │ 44.71 ms      │ 44.69 ms      │ 100     │ 100
│             119.1 Kitem/s │ 101.8 Kitem/s │ 111.8 Kitem/s │ 111.8 Kitem/s │         │
╰─ spans                    │               │               │               │         │
   ├─ 1       55.5 ms       │ 76.2 ms       │ 58.85 ms      │ 60.54 ms      │ 100     │ 100
   │          90.08 Kitem/s │ 65.61 Kitem/s │ 84.95 Kitem/s │ 82.57 Kitem/s │         │
   ├─ 4       31.67 ms      │ 56.38 ms      │ 34.49 ms      │ 35.72 ms      │ 100     │ 100
   │          157.8 Kitem/s │ 88.68 Kitem/s │ 144.9 Kitem/s │ 139.9 Kitem/s │         │
   ╰─ 16      22.08 ms      │ 72.64 ms      │ 24.25 ms      │ 25.54 ms      │ 100     │ 100
              226.3 Kitem/s │ 68.82 Kitem/s │ 206.1 Kitem/s │ 195.6 Kitem/s │         │
```